# openapi.yaml
openapi: 3.0.3
info:
  title: LeadOS API
  version: 0.1.0
  description: >
    AI-Powered CRM & Social Lead Automation Platform. Multi-tenant. Most routes
    require a Bearer JWT and an `X-Org-Id` header identifying the organization
    (tenant isolation, BRD §20). Import this file directly into Postman
    (File → Import) or any OpenAPI client.
servers:
  - url: http://localhost:4000
    description: Local
  - url: https://leados-api.onrender.com
    description: Production (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: X-Org-Id
      in: header
      required: true
      schema: { type: string }
      description: Organization id (tenant scope, BRD §20)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/contacts/export.csv:
    get: { tags: [Contacts], summary: Export contacts as CSV, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '200': { description: CSV } } }
  /api/v1/contacts/import:
    post: { tags: [Contacts], summary: Bulk import contacts, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '201': { description: Imported } } }
  /api/v1/contacts/bulk:
    post: { tags: [Contacts], summary: Bulk delete contacts, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '200': { description: '{ affected }' } } }
...
  /api/v1/contacts:
    get: { tags: [Contacts], summary: List contacts, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '200': { description: OK } } }
    post: { tags: [Contacts], summary: Create contact, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '201': { description: Created } } }
  /api/v1/contacts/{id}/notes:
    post:
      tags: [Contacts]
      summary: Add a note to a contact's timeline
      parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { note: { type: string } } } } } }
      responses: { '201': { description: Logged } }
  /api/v1/contacts/{id}/customer360:
    get:
      tags: [Contacts]
      summary: Customer 360 profile (BRD §10.7)
      parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
...